requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
azure-cosmos
